add tests and small tidy-ups
